+++ app/Repositories/VariantRepository.php
@@ -8,35 +8,32 @@ use Illuminate\Support\Facades\Validator;
 
 class VariantRepository
 {
-    public function index($filters)
+    public function index($filters, $companyId)
     {
         $variant = Variant::with([]);
         if (!empty($filters['title'])) {
             $variant = $variant->where('title', 'LIKE', '%' . $filters['title'] . '%');
         }
-        if (!empty($filters['company_id'])) {
-            $variant = $variant->where('company_id', $filters['company_id']);
-        }
+        $variant = $variant->where('company_id', $companyId);
         $variant = $variant->orderBy('id', 'desc')->paginate(25);
         return $variant;
     }
 
-    public function save($data)
+    public function save($data, $companyId)
     {
         try {
             $validator = Validator::make($data, [
-                'company_id' => 'required',
                 'title' => 'required',
                 'price' => 'required',
                 'image' => 'required',
             ]);
             if ($validator->fails()) return resultFunction('Err code VR-S: validation err ' . $validator->errors());
 
-            $company = Company::find($data['company_id']);
+            $company = Company::find($companyId);
             if (!$company) return resultFunction('Err code VR-S: company not found');
 
             $variant = new Variant();
-            $variant->company_id = $data['company_id'];
+            $variant->company_id = $company->id;
             $variant->title = $data['title'];
             $variant->price = $data['price'];
             $variant->image = json_encode($data['image']);
@@ -48,10 +45,12 @@ class VariantRepository
         }
     }
 
-    public function delete($id) {
+    public function delete($id, $companyId) {
         try {
             $variant =  Variant::find($id);
             if (!$variant) return resultFunction('Err VR-D: product category not found');
+
+            if ($variant->company_id != $companyId) return resultFunction('Err VR-D: variant not found');
             $variant->delete();
 
             return resultFunction("Success to delete variant", true);
